server/requirements.txt
Flask>=3.0.3
flask_sqlalchemy>=3.1.1
SQLAlchemy>=2.0.31
PyMySQL>=1.1.1
flask_cors>=4.0.1
pytest>=8.2.2
python-dotenv>=1.0.1